[C/Wrapper] Separate the aeron_uri parser into client and driver parts.  Add support for a print function for the aeron_uri.  Implement aeron_subscription_try_resolve_endpoint_port using aeron_uri parser and new print function.  Have wrapper use the C API function for Subscription::tryResolveEndpointPort.
