+++ aeron-client/src/main/c/aeronc.h
@@ -32,6 +32,7 @@ extern "C"
 #define AERON_CLIENT_ERROR_CLIENT_TIMEOUT (-1001)
 #define AERON_CLIENT_ERROR_CONDUCTOR_SERVICE_TIMEOUT (-1002)
 #define AERON_CLIENT_ERROR_BUFFER_FULL (-1003)
+#define AERON_CLIENT_MAX_LOCAL_ADDRESS_STR_LEN (64)
 
 typedef struct aeron_context_stct aeron_context_t;
 typedef struct aeron_stct aeron_t;
@@ -54,14 +55,14 @@ typedef struct aeron_header_values_frame_stct
     int32_t term_id;
     int64_t reserved_value;
 }
-aeron_header_values_frame_t;
+    aeron_header_values_frame_t;
 
 typedef struct aeron_header_values_stct
 {
     aeron_header_values_frame_t frame;
     int32_t initial_term_id;
 }
-aeron_header_values_t;
+    aeron_header_values_t;
 #pragma pack(pop)
 
 typedef struct aeron_subscription_stct aeron_subscription_t;
@@ -92,24 +93,29 @@ typedef struct aeron_controlled_fragment_assembler_stct aeron_controlled_fragmen
 #define AERON_DIR_ENV_VAR "AERON_DIR"
 
 int aeron_context_set_dir(aeron_context_t *context, const char *value);
+
 const char *aeron_context_get_dir(aeron_context_t *context);
 
 #define AERON_DRIVER_TIMEOUT_ENV_VAR "AERON_DRIVER_TIMEOUT"
 
 int aeron_context_set_driver_timeout_ms(aeron_context_t *context, uint64_t value);
+
 uint64_t aeron_context_get_driver_timeout_ms(aeron_context_t *context);
 
 int aeron_context_set_keepalive_interval_ns(aeron_context_t *context, uint64_t value);
+
 uint64_t aeron_context_get_keepalive_interval_ns(aeron_context_t *context);
 
 #define AERON_CLIENT_RESOURCE_LINGER_DURATION_ENV_VAR "AERON_CLIENT_RESOURCE_LINGER_DURATION"
 
 int aeron_context_set_resource_linger_duration_ns(aeron_context_t *context, uint64_t value);
+
 uint64_t aeron_context_get_resource_linger_duration_ns(aeron_context_t *context);
 
 #define AERON_CLIENT_PRE_TOUCH_MAPPED_MEMORY_ENV_VAR "AERON_CLIENT_PRE_TOUCH_MAPPED_MEMORY"
 
 int aeron_context_set_pre_touch_mapped_memory(aeron_context_t *context, bool value);
+
 bool aeron_context_get_pre_touch_mapped_memory(aeron_context_t *context);
 
 /**
@@ -123,7 +129,9 @@ typedef void (*aeron_error_handler_t)(void *clientd, int errcode, const char *me
 typedef void (*aeron_notification_t)(void *clientd);
 
 int aeron_context_set_error_handler(aeron_context_t *context, aeron_error_handler_t handler, void *clientd);
+
 aeron_error_handler_t aeron_context_get_error_handler(aeron_context_t *context);
+
 void *aeron_context_get_error_handler_clientd(aeron_context_t *context);
 
 /**
@@ -148,12 +156,16 @@ typedef void (*aeron_on_new_publication_t)(
     int64_t correlation_id);
 
 int aeron_context_set_on_new_publication(aeron_context_t *context, aeron_on_new_publication_t handler, void *clientd);
+
 aeron_on_new_publication_t aeron_context_get_on_new_publication(aeron_context_t *context);
+
 void *aeron_context_get_on_new_publication_clientd(aeron_context_t *context);
 
 int aeron_context_set_on_new_exclusive_publication(
     aeron_context_t *context, aeron_on_new_publication_t handler, void *clientd);
+
 aeron_on_new_publication_t aeron_context_get_on_new_exclusive_publication(aeron_context_t *context);
+
 void *aeron_context_get_on_new_exclusive_publication_clientd(aeron_context_t *context);
 
 /**
@@ -178,7 +190,9 @@ typedef void (*aeron_on_new_subscription_t)(
 
 int aeron_context_set_on_new_subscription(
     aeron_context_t *context, aeron_on_new_subscription_t handler, void *clientd);
+
 aeron_on_new_subscription_t aeron_context_get_on_new_subscription(aeron_context_t *context);
+
 void *aeron_context_get_on_new_subscription_clientd(aeron_context_t *context);
 
 /**
@@ -213,7 +227,9 @@ typedef void (*aeron_on_available_counter_t)(
 
 int aeron_context_set_on_available_counter(
     aeron_context_t *context, aeron_on_available_counter_t handler, void *clientd);
+
 aeron_on_available_counter_t aeron_context_get_on_available_counter(aeron_context_t *context);
+
 void *aeron_context_get_on_available_counter_clientd(aeron_context_t *context);
 
 /**
@@ -229,7 +245,9 @@ typedef void (*aeron_on_unavailable_counter_t)(
 
 int aeron_context_set_on_unavailable_counter(
     aeron_context_t *context, aeron_on_unavailable_counter_t handler, void *clientd);
+
 aeron_on_unavailable_counter_t aeron_context_get_on_unavailable_counter(aeron_context_t *context);
+
 void *aeron_context_get_on_unavailable_counter_clientd(aeron_context_t *context);
 
 /**
@@ -241,13 +259,16 @@ typedef void (*aeron_on_close_client_t)(void *clientd);
 
 int aeron_context_set_on_close_client(
     aeron_context_t *context, aeron_on_close_client_t handler, void *clientd);
+
 aeron_on_close_client_t aeron_context_get_on_close_client(aeron_context_t *context);
+
 void *aeron_context_get_on_close_client_clientd(aeron_context_t *context);
 
 /**
  * Whether to use an invoker to control the conductor agent or spawn a thread.
  */
 int aeron_context_set_use_conductor_agent_invoker(aeron_context_t *context, bool value);
+
 bool aeron_context_get_use_conductor_agent_invoker(aeron_context_t *context);
 
 /**
@@ -259,7 +280,9 @@ typedef void (*aeron_agent_on_start_func_t)(void *state, const char *role_name);
 
 int aeron_context_set_agent_on_start_function(
     aeron_context_t *context, aeron_agent_on_start_func_t value, void *state);
+
 aeron_agent_on_start_func_t aeron_context_get_agent_on_start_function(aeron_context_t *context);
+
 void *aeron_context_get_agent_on_start_state(aeron_context_t *context);
 
 /**
@@ -492,21 +515,21 @@ typedef struct aeron_on_available_counter_pair_stct
     aeron_on_available_counter_t handler;
     void *clientd;
 }
-aeron_on_available_counter_pair_t;
+    aeron_on_available_counter_pair_t;
 
 typedef struct aeron_on_unavailable_counter_pair_stct
 {
     aeron_on_unavailable_counter_t handler;
     void *clientd;
 }
-aeron_on_unavailable_counter_pair_t;
+    aeron_on_unavailable_counter_pair_t;
 
 typedef struct aeron_on_close_client_pair_stct
 {
     aeron_on_close_client_t handler;
     void *clientd;
 }
-aeron_on_close_client_pair_t;
+    aeron_on_close_client_pair_t;
 
 /**
  * Add a handler to be called when a new counter becomes available.
@@ -590,7 +613,7 @@ typedef struct aeron_counter_value_descriptor_stct
     int64_t owner_id;
     uint8_t pad1[(2 * AERON_COUNTER_CACHE_LINE_LENGTH) - (3 * sizeof(int64_t))];
 }
-aeron_counter_value_descriptor_t;
+    aeron_counter_value_descriptor_t;
 
 typedef struct aeron_counter_metadata_descriptor_stct
 {
@@ -601,7 +624,7 @@ typedef struct aeron_counter_metadata_descriptor_stct
     int32_t label_length;
     uint8_t label[(6 * AERON_COUNTER_CACHE_LINE_LENGTH) - sizeof(int32_t)];
 }
-aeron_counter_metadata_descriptor_t;
+    aeron_counter_metadata_descriptor_t;
 #pragma pack(pop)
 
 
@@ -637,7 +660,7 @@ typedef struct aeron_counters_reader_buffers_stct
     size_t values_length;
     size_t metadata_length;
 }
-aeron_counters_reader_buffers_t;
+    aeron_counters_reader_buffers_t;
 
 /**
  * Get buffer pointers and lengths for the counters reader.
@@ -817,7 +840,7 @@ typedef struct aeron_iovec_stct
     uint8_t *iov_base;
     size_t iov_len;
 }
-aeron_iovec_t;
+    aeron_iovec_t;
 #else
 typedef struct iov aeron_iovec_t;
 #endif
@@ -831,7 +854,7 @@ typedef struct aeron_buffer_claim_stct
     uint8_t *data;
     size_t length;
 }
-aeron_buffer_claim_t;
+    aeron_buffer_claim_t;
 
 /**
  * Commit the given buffer_claim as a complete message available for consumption.
@@ -929,7 +952,7 @@ typedef struct aeron_publication_constants_stct
      */
     int32_t channel_status_indicator_id;
 }
-aeron_publication_constants_t;
+    aeron_publication_constants_t;
 
 /**
  * Non-blocking publish of a buffer containing a message.
@@ -1385,7 +1408,7 @@ typedef enum aeron_controlled_fragment_handler_action_en
      */
     AERON_ACTION_CONTINUE = 4
 }
-aeron_controlled_fragment_handler_action_t;
+    aeron_controlled_fragment_handler_action_t;
 
 /**
  * Callback for handling fragments of data being read from a log.
@@ -1662,8 +1685,9 @@ int aeron_subscription_close(
  * multi-destination subscription.  Addresses will a string representation in numeric form.  IPv6 addresses will be
  * surrounded by '[' and ']' so that the ':' that separate the parts are distinguishable from the port delimiter.
  * E.g. [fe80::7552:c06e:6bf4:4160]:12345.  As of writing the maximum length for a formatted address is 54 bytes
- * including the NULL terminator.  Returned strings will be NULL terminated.  If the buffer to hold the address can
- * not hold enough of the message it will be truncated and the last character will be null.
+ * including the NULL terminator.  AERON_CLIENT_MAX_LOCAL_ADDRESS_STR_LEN is defined to provide enough space to fit the
+ * returned string. Returned strings will be NULL terminated.  If the buffer to hold the address can not hold enough
+ * of the message it will be truncated and the last character will be null.
  *
  * @param subscription to query
  * @param address_vec to hold the received addresses
@@ -1684,6 +1708,12 @@ int aeron_subscription_local_sockaddrs(
  * @return -1 on error, 0 if address not found, 1 if address is found.
  */
 int aeron_subscription_resolved_endpoint(aeron_subscription_t *subscription, char *address, size_t address_len);
+
+
+int aeron_subscription_try_resolve_channel_endpoint_port(
+    aeron_subscription_t *subscription,
+    char *address,
+    size_t address_len);
 
 /**
  * Image Functions
@@ -1756,7 +1786,7 @@ typedef struct aeron_image_constants_stct
      */
     int32_t subscriber_position_id;
 }
-aeron_image_constants_t;
+    aeron_image_constants_t;
 
 /**
  * Fill in a structure with the constants in use by a image.
@@ -2079,7 +2109,7 @@ typedef struct aeron_counter_constants_stct
      */
     int32_t counter_id;
 }
-aeron_counter_constants_t;
+    aeron_counter_constants_t;
 
 /**
  * Fill in a structure with the constants in use by a counter.
