+++ aeron-client/src/test/c/CMakeLists.txt
@@ -42,3 +42,4 @@ aeron_c_client_test(image_test aeron_image_test.cpp)
 aeron_c_client_test(fragment_assembler_test aeron_fragment_assembler_test.cpp)
 aeron_c_client_test(array_to_ptr_hash_map_test collections/aeron_array_to_ptr_hash_map_test.cpp)
 aeron_c_client_test(aeron_fileutil_test util/aeron_fileutil_test.cpp)
+aeron_c_client_test(uri_test aeron_uri_test.cpp)
