+++ aeron-driver/src/main/c/CMakeLists.txt
@@ -153,6 +153,7 @@ SET(C_CLIENT_SOURCE
     ${AERON_C_CLIENT_SOURCE_PATH}/util/aeron_parse_util.c
     ${AERON_C_CLIENT_SOURCE_PATH}/util/aeron_properties_util.c
     ${AERON_C_CLIENT_SOURCE_PATH}/util/aeron_strutil.c
+    ${AERON_C_CLIENT_SOURCE_PATH}/uri/aeron_uri.c
     ${AERON_C_CLIENT_SOURCE_PATH}/aeron_agent.c
     ${AERON_C_CLIENT_SOURCE_PATH}/aeron_alloc.c
     ${AERON_C_CLIENT_SOURCE_PATH}/aeron_client.c
@@ -250,7 +251,7 @@ SET(SOURCE
     media/aeron_udp_destination_tracker.c
     media/aeron_udp_transport_poller.c
     reports/aeron_loss_reporter.c
-    uri/aeron_uri.c
+    uri/aeron_uri_driver.c
     aeron_congestion_control.c
     aeron_csv_table_name_resolver.c
     aeron_data_packet_dispatcher.c
@@ -290,7 +291,7 @@ SET(HEADERS
     media/aeron_udp_destination_tracker.h
     media/aeron_udp_transport_poller.h
     reports/aeron_loss_reporter.h
-    uri/aeron_uri.h
+    uri/aeron_uri_driver.h
     aeron_congestion_control.h
     aeron_data_packet_dispatcher.h
     aeron_driver.h
