+++ aeron-driver/src/main/c/aeron_ipc_publication.c
@@ -32,7 +32,7 @@ int aeron_ipc_publication_create(
     aeron_position_t *pub_pos_position,
     aeron_position_t *pub_lmt_position,
     int32_t initial_term_id,
-    aeron_uri_publication_params_t *params,
+    aeron_driver_uri_publication_params_t *params,
     bool is_exclusive,
     aeron_system_counters_t *system_counters,
     const size_t channel_length,
