+++ aeron-driver/src/test/c/CMakeLists.txt
@@ -52,7 +52,7 @@ aeron_driver_test(driver_conductor_counter_test aeron_driver_conductor_counter_t
 aeron_driver_test(driver_conductor_clock_test aeron_driver_conductor_clock_test.cpp)
 aeron_driver_test(spsc_queue_test aeron_spsc_concurrent_array_queue_test.cpp)
 aeron_driver_test(mpsc_queue_test aeron_mpsc_concurrent_array_queue_test.cpp)
-aeron_driver_test(uri_test aeron_uri_test.cpp)
+aeron_driver_test(uri_driver_test aeron_uri_driver_test.cpp)
 aeron_driver_test(udp_channel_test aeron_udp_channel_test.cpp)
 aeron_driver_test(int64_to_ptr_hash_map_test collections/aeron_int64_to_ptr_hash_map_test.cpp)
 aeron_driver_test(int64_counter_map_test collections/aeron_int64_counter_map_test.cpp)
