+++ aeron-driver/src/test/c/aeron_uri_driver_test.cpp
@@ -29,10 +29,10 @@ extern "C"
 #include "aeron_name_resolver.h"
 }
 
-class UriTest : public testing::Test
+class UriDriverTest : public testing::Test
 {
 public:
-    UriTest()
+    UriDriverTest()
     {
         if (aeron_driver_context_init(&m_context) < 0)
         {
@@ -42,7 +42,7 @@ public:
         m_conductor.context = m_context;
     }
 
-    ~UriTest() override
+    ~UriDriverTest() override
     {
         aeron_uri_close(&m_uri);
         aeron_driver_context_close(m_context);
@@ -56,83 +56,7 @@ protected:
 
 #define AERON_URI_PARSE(uri_str, uri) aeron_uri_parse(strlen(uri_str), uri_str, uri)
 
-TEST_F(UriTest, shouldNotParseInvalidUriScheme)
-{
-    EXPECT_EQ(AERON_URI_PARSE("aaron", &m_uri), -1);
-    EXPECT_EQ(AERON_URI_PARSE("aeron:", &m_uri), -1);
-    EXPECT_EQ(AERON_URI_PARSE("aron:", &m_uri), -1);
-    EXPECT_EQ(AERON_URI_PARSE(":aeron", &m_uri), -1);
-    EXPECT_EQ(AERON_URI_PARSE("aeron:udp:", &m_uri), -1);
-}
-
-TEST_F(UriTest, shouldNotParseUnknownUriTransport)
-{
-    EXPECT_EQ(AERON_URI_PARSE("aeron:tcp", &m_uri), -1);
-    EXPECT_EQ(AERON_URI_PARSE("aeron:sctp", &m_uri), -1);
-    EXPECT_EQ(AERON_URI_PARSE("aeron:udp", &m_uri), -1);
-    EXPECT_EQ(AERON_URI_PARSE("aeron:ipcsdfgfdhfgf", &m_uri), -1);
-}
-
-TEST_F(UriTest, shouldRejectWithMissingQuerySeparatorWhenFollowedWithParams)
-{
-    EXPECT_EQ(AERON_URI_PARSE("aeron:ipc|sparse=true", &m_uri), -1);
-}
-
-TEST_F(UriTest, shouldRejectWithInvalidParams)
-{
-    EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=localhost:4652|-~@{]|=??#s!£$%====", &m_uri), -1);
-    EXPECT_EQ(AERON_URI_PARSE("aeron:udp?add|ress=224.10.9.8", &m_uri), -1);
-}
-
-TEST_F(UriTest, shouldParseKnownUriTransportWithoutParamsIpcNoSeparator)
-{
-    EXPECT_EQ(AERON_URI_PARSE("aeron:ipc", &m_uri), 0);
-    ASSERT_EQ(m_uri.type, AERON_URI_IPC);
-    EXPECT_EQ(m_uri.params.ipc.additional_params.length, 0u);
-}
-
-TEST_F(UriTest, shouldParseKnownUriTransportWithoutParamsUdpWithSeparator)
-{
-    EXPECT_EQ(AERON_URI_PARSE("aeron:udp?", &m_uri), 0);
-    EXPECT_EQ(m_uri.type, AERON_URI_UDP);
-}
-
-TEST_F(UriTest, shouldParseKnownUriTransportWithoutParamsIpcWithSeparator)
-{
-    EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?", &m_uri), 0);
-    EXPECT_EQ(m_uri.type, AERON_URI_IPC);
-    EXPECT_EQ(m_uri.params.ipc.additional_params.length, 0u);
-}
-
-TEST_F(UriTest, shouldParseWithSingleParamUdpEndpoint)
-{
-    EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8", &m_uri), 0);
-    ASSERT_EQ(m_uri.type, AERON_URI_UDP);
-    EXPECT_EQ(std::string(m_uri.params.udp.endpoint), "224.10.9.8");
-    EXPECT_EQ(m_uri.params.udp.additional_params.length, 0u);
-}
-
-TEST_F(UriTest, shouldParseWithSingleParamUdpValueWithEmbeddedEquals)
-{
-    EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.1=0.9.8", &m_uri), 0);
-    ASSERT_EQ(m_uri.type, AERON_URI_UDP);
-    EXPECT_EQ(std::string(m_uri.params.udp.endpoint), "224.1=0.9.8");
-    EXPECT_EQ(m_uri.params.udp.additional_params.length, 0u);
-}
-
-TEST_F(UriTest, shouldParseWithMultipleParams)
-{
-    EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8|port=4567|interface=192.168.0.3|ttl=16", &m_uri), 0);
-    ASSERT_EQ(m_uri.type, AERON_URI_UDP);
-    EXPECT_EQ(std::string(m_uri.params.udp.endpoint), "224.10.9.8");
-    EXPECT_EQ(std::string(m_uri.params.udp.bind_interface), "192.168.0.3");
-    EXPECT_EQ(std::string(m_uri.params.udp.ttl), "16");
-    EXPECT_EQ(m_uri.params.udp.additional_params.length, 1u);
-    EXPECT_EQ(std::string(m_uri.params.udp.additional_params.array[0].key), "port");
-    EXPECT_EQ(std::string(m_uri.params.udp.additional_params.array[0].value), "4567");
-}
-
-TEST_F(UriTest, shouldParseCongestionControlParam)
+TEST_F(UriDriverTest, shouldParseCongestionControlParam)
 {
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8|cc=static", &m_uri), 0);
     ASSERT_EQ(m_uri.type, AERON_URI_UDP);
@@ -142,230 +66,230 @@ TEST_F(UriTest, shouldParseCongestionControlParam)
     EXPECT_EQ(std::string(m_uri.params.udp.additional_params.array[0].value), AERON_STATICWINDOWCONGESTIONCONTROL_CC_PARAM_VALUE);
 }
 
-TEST_F(UriTest, shouldParseNoPublicationParams)
+TEST_F(UriDriverTest, shouldParseNoPublicationParams)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), 0);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), 0);
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), 0);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), 0);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamLingerTimeout)
+TEST_F(UriDriverTest, shouldParsePublicationParamLingerTimeout)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8|linger=7777", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), 0);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), 0);
     EXPECT_EQ(params.linger_timeout_ns, 7777u);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamUdpTermLength)
+TEST_F(UriDriverTest, shouldParsePublicationParamUdpTermLength)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8|term-length=131072", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), 0);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), 0);
     EXPECT_EQ(params.term_length, 131072u);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamIpcTermLength)
+TEST_F(UriDriverTest, shouldParsePublicationParamIpcTermLength)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?term-length=262144", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), 0);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), 0);
     EXPECT_EQ(params.term_length, 262144u);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamIpcTermLengthOddValue)
+TEST_F(UriDriverTest, shouldParsePublicationParamIpcTermLengthOddValue)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?term-length=262143", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), -1);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), -1);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamIpcTermLength32K)
+TEST_F(UriDriverTest, shouldParsePublicationParamIpcTermLength32K)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?term-length=32768", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), -1);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), -1);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamIpcTermLength2T)
+TEST_F(UriDriverTest, shouldParsePublicationParamIpcTermLength2T)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?term-length=2147483648", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), -1);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), -1);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamUdpEndpointAndMtuLength)
+TEST_F(UriDriverTest, shouldParsePublicationParamUdpEndpointAndMtuLength)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8|mtu=18432", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), 0);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), 0);
     EXPECT_EQ(params.mtu_length, 18432u);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamIpcMtuLength32K)
+TEST_F(UriDriverTest, shouldParsePublicationParamIpcMtuLength32K)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?mtu=32768", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), 0);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), 0);
     EXPECT_EQ(params.mtu_length, 32768u);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamIpcMtuLengthNonPowerOf2)
+TEST_F(UriDriverTest, shouldParsePublicationParamIpcMtuLengthNonPowerOf2)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?mtu=66560", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), -1);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), -1);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamIpcMtuLengthLessThanHeaderLength)
+TEST_F(UriDriverTest, shouldParsePublicationParamIpcMtuLengthLessThanHeaderLength)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?mtu=10", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), -1);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), -1);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamIpcMtuLengthNotMultipleOfFrameAlignment)
+TEST_F(UriDriverTest, shouldParsePublicationParamIpcMtuLengthNotMultipleOfFrameAlignment)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?mtu=255", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), -1);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), -1);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamIpcSparse)
+TEST_F(UriDriverTest, shouldParsePublicationParamIpcSparse)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?sparse=true", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), 0);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), 0);
     EXPECT_EQ(params.is_sparse, true);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamsForReplayUdp)
+TEST_F(UriDriverTest, shouldParsePublicationParamsForReplayUdp)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8|init-term-id=120|term-id=127|term-offset=64", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, true), 0) << aeron_errmsg();
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, true), 0) << aeron_errmsg();
     EXPECT_EQ(params.has_position, true);
     EXPECT_EQ(params.initial_term_id, 120l);
     EXPECT_EQ(params.term_id, 127l);
     EXPECT_EQ(params.term_offset, 64u);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamsForReplayIpc)
+TEST_F(UriDriverTest, shouldParsePublicationParamsForReplayIpc)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?init-term-id=250|term-id=257|term-offset=128", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, true), 0) << aeron_errmsg();
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, true), 0) << aeron_errmsg();
     EXPECT_EQ(params.has_position, true);
     EXPECT_EQ(params.initial_term_id, 250l);
     EXPECT_EQ(params.term_id, 257l);
     EXPECT_EQ(params.term_offset, 128u);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamsForReplayIpcNegativeTermIds)
+TEST_F(UriDriverTest, shouldParsePublicationParamsForReplayIpcNegativeTermIds)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?init-term-id=-257|term-id=-250|term-offset=128", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, true), 0) << aeron_errmsg();
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, true), 0) << aeron_errmsg();
     EXPECT_EQ(params.initial_term_id, -257l);
     EXPECT_EQ(params.term_id, -250l);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamsForReplayIpcOddTermOffset)
+TEST_F(UriDriverTest, shouldParsePublicationParamsForReplayIpcOddTermOffset)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:ipc?init-term-id=-257|term-id=-250|term-offset=127", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, true), -1);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, true), -1);
 }
 
-TEST_F(UriTest, shouldParsePublicationParamsForReplayIpcTermOffsetBeyondTermLength)
+TEST_F(UriDriverTest, shouldParsePublicationParamsForReplayIpcTermOffsetBeyondTermLength)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE(
         "aeron:ipc?term-length=65536|init-term-id=-257|term-id=-250|term-offset=65537", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, true), -1);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, true), -1);
 }
 
-TEST_F(UriTest, shouldErrorParsingTooLargeTermIdRange)
+TEST_F(UriDriverTest, shouldErrorParsingTooLargeTermIdRange)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE(
         "aeron:ipc?term-length=65536|init-term-id=-2147483648|term-id=0|term-offset=0", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, true), -1);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, true), -1);
     EXPECT_THAT(std::string(aeron_errmsg()), ::testing::StartsWith("Param difference greater than 2^31 - 1"));
 }
 
-TEST_F(UriTest, shouldParsePublicationSessionId)
+TEST_F(UriDriverTest, shouldParsePublicationSessionId)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8|session-id=1001", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), 0);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), 0);
     EXPECT_EQ(params.has_session_id, true);
     EXPECT_EQ(params.session_id, 1001);
 }
 
-TEST_F(UriTest, shouldErrorParsingNonNumericSessionId)
+TEST_F(UriDriverTest, shouldErrorParsingNonNumericSessionId)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8|session-id=foobar", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), -1);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), -1);
 }
 
-TEST_F(UriTest, shouldErrorParsingOutOfRangeSessionId)
+TEST_F(UriDriverTest, shouldErrorParsingOutOfRangeSessionId)
 {
-    aeron_uri_publication_params_t params;
+    aeron_driver_uri_publication_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8|session-id=2147483648", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_publication_params(&m_uri, &params, &m_conductor, false), -1);
+    EXPECT_EQ(aeron_uri_driver_publication_params(&m_uri, &params, &m_conductor, false), -1);
 }
 
-TEST_F(UriTest, shouldParseSubscriptionParamReliable)
+TEST_F(UriDriverTest, shouldParseSubscriptionParamReliable)
 {
     aeron_uri_subscription_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8|reliable=false", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_subscription_params(&m_uri, &params, &m_conductor), 0);
+    EXPECT_EQ(aeron_uri_driver_subscription_params(&m_uri, &params, &m_conductor), 0);
     EXPECT_EQ(params.is_reliable, false);
 }
 
-TEST_F(UriTest, shouldParseSubscriptionParamReliableDefault)
+TEST_F(UriDriverTest, shouldParseSubscriptionParamReliableDefault)
 {
     aeron_uri_subscription_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_subscription_params(&m_uri, &params, &m_conductor), 0);
+    EXPECT_EQ(aeron_uri_driver_subscription_params(&m_uri, &params, &m_conductor), 0);
     EXPECT_EQ(params.is_reliable, true);
 }
 
-TEST_F(UriTest, shouldParseSubscriptionSessionId)
+TEST_F(UriDriverTest, shouldParseSubscriptionSessionId)
 {
     aeron_uri_subscription_params_t params;
 
     EXPECT_EQ(AERON_URI_PARSE("aeron:udp?endpoint=224.10.9.8|session-id=1001", &m_uri), 0);
-    EXPECT_EQ(aeron_uri_subscription_params(&m_uri, &params, &m_conductor), 0);
+    EXPECT_EQ(aeron_uri_driver_subscription_params(&m_uri, &params, &m_conductor), 0);
     EXPECT_EQ(params.has_session_id, true);
     EXPECT_EQ(params.session_id, 1001);
 }
@@ -634,136 +558,4 @@ TEST_F(UriResolverTest, shouldCalculateIpv6PrefixlenFromNetmask)
     EXPECT_EQ(ipv6_prefixlen("FFFF:FFFF:FFFF:FFFF:FFFF:FFFF:FF00::"), 104u);
     EXPECT_EQ(ipv6_prefixlen("FFFF:FF80::"), 25u);
     EXPECT_EQ(ipv6_prefixlen("0000:0000:0000:0000:0000:0000:0000:0000"), 0u);
-}
-
-/*
- * WARNING: single threaded only due to global lookup func usage
- */
-
-struct ifaddrs *global_ifaddrs = nullptr;
-
-class UriLookupTest : public testing::Test
-{
-public:
-    UriLookupTest()
-    {
-        aeron_set_getifaddrs(UriLookupTest::getifaddrs, UriLookupTest::freeifaddrs);
-    }
-
-    static void add_ifaddr(
-        int family, const char *name, const char *addr_str, const char *netmask_str, unsigned int flags)
-    {
-        struct ifaddrs *entry = new struct ifaddrs;
-        void *addr, *netmask;
-
-        if (family == AF_INET6)
-        {
-            struct sockaddr_in6 *a = new struct sockaddr_in6;
-            addr = &a->sin6_addr;
-            a->sin6_family = AF_INET6;
-            entry->ifa_addr = reinterpret_cast<struct sockaddr *>(a);
-
-            struct sockaddr_in6 *b = new struct sockaddr_in6;
-            netmask = &b->sin6_addr;
-            b->sin6_family = AF_INET6;
-            entry->ifa_netmask = reinterpret_cast<struct sockaddr *>(b);
-        }
-        else
-        {
-            struct sockaddr_in *a = new struct sockaddr_in;
-            addr = &a->sin_addr;
-            a->sin_family = AF_INET;
-            entry->ifa_addr = reinterpret_cast<struct sockaddr *>(a);
-
-            struct sockaddr_in *b = new struct sockaddr_in;
-            netmask = &b->sin_addr;
-            b->sin_family = AF_INET;
-            entry->ifa_netmask = reinterpret_cast<struct sockaddr *>(b);
-        }
-
-        if (inet_pton(family, addr_str, addr) != 1 || inet_pton(family, netmask_str, netmask) != 1)
-        {
-            throw std::runtime_error("could not convert address");
-        }
-
-        entry->ifa_name = ::strdup(name);
-        entry->ifa_flags = flags;
-        entry->ifa_next = global_ifaddrs;
-        global_ifaddrs = entry;
-    }
-
-    static void initialize_ifaddrs()
-    {
-        if (nullptr == global_ifaddrs)
-        {
-            add_ifaddr(AF_INET, "lo0", "127.0.0.1", "255.0.0.0", IFF_MULTICAST | IFF_UP | IFF_LOOPBACK);
-            add_ifaddr(AF_INET, "eth0:0", "192.168.0.20", "255.255.255.0", IFF_MULTICAST | IFF_UP);
-            add_ifaddr(AF_INET, "eth0:1", "192.168.1.21", "255.255.255.0", IFF_MULTICAST | IFF_UP);
-            add_ifaddr(AF_INET6, "eth1:0", "ee80:0:0:0001:0:0:0:1", "FFFF:FFFF:FFFF:FFFF::", IFF_MULTICAST | IFF_UP);
-            add_ifaddr(AF_INET6, "eth1:3", "fe80:1:abcd:0:0:0:0:1", "FFFF:FFFF:FFFF::", IFF_MULTICAST | IFF_UP);
-            add_ifaddr(AF_INET6, "eth1:1", "fe80:0:0:0:0:0:0:1", "FFFF::", IFF_MULTICAST | IFF_UP);
-            add_ifaddr(AF_INET6, "eth1:2", "fe80:1:0:0:0:0:0:1", "FFFF:FFFF::", IFF_MULTICAST | IFF_UP);
-        }
-    }
-
-    static int getifaddrs(struct ifaddrs **ifaddrs)
-    {
-        initialize_ifaddrs();
-        *ifaddrs = global_ifaddrs;
-        return 0;
-    }
-
-    static void freeifaddrs(struct ifaddrs *)
-    {
-    }
-
-protected:
-};
-
-TEST_F(UriLookupTest, shouldFindIpv4Loopback)
-{
-    char buffer[AERON_MAX_PATH];
-    struct sockaddr_storage addr;
-    struct sockaddr_in *addr_in = (struct sockaddr_in *)&addr;
-    unsigned int if_index;
-
-    ASSERT_EQ(aeron_find_interface("127.0.0.0/16", (struct sockaddr_storage *)&addr, &if_index), 0);
-    EXPECT_EQ(addr_in->sin_family, AF_INET);
-    EXPECT_STREQ(inet_ntop(AF_INET, &addr_in->sin_addr, buffer, sizeof(buffer)), "127.0.0.1");
-}
-
-TEST_F(UriLookupTest, shouldFindIpv4LoopbackAsLocalhost)
-{
-    char buffer[AERON_MAX_PATH];
-    struct sockaddr_storage addr;
-    struct sockaddr_in *addr_in = (struct sockaddr_in *)&addr;
-    unsigned int if_index;
-
-    ASSERT_EQ(aeron_find_interface("localhost:40123", (struct sockaddr_storage *)&addr, &if_index), 0);
-    EXPECT_EQ(addr_in->sin_family, AF_INET);
-    EXPECT_STREQ(inet_ntop(AF_INET, &addr_in->sin_addr, buffer, sizeof(buffer)), "127.0.0.1");
-    EXPECT_EQ(addr_in->sin_port, htons(40123));
-}
-
-TEST_F(UriLookupTest, shouldFindIpv6)
-{
-    char buffer[AERON_MAX_PATH];
-    struct sockaddr_storage addr;
-    struct sockaddr_in6 *addr_in6 = (struct sockaddr_in6 *)&addr;
-    unsigned int if_index;
-
-    ASSERT_EQ(aeron_find_interface("[fe80:0:0:0:0:0:0:0]/16", (struct sockaddr_storage *)&addr, &if_index), 0);
-    EXPECT_EQ(addr_in6->sin6_family, AF_INET6);
-    EXPECT_STREQ(inet_ntop(AF_INET6, &addr_in6->sin6_addr, buffer, sizeof(buffer)), "fe80:1:abcd::1");
-}
-
-TEST_F(UriLookupTest, shouldNotFindUnknown)
-{
-    struct sockaddr_storage addr;
-    unsigned int if_index;
-
-    ASSERT_EQ(aeron_find_interface("[fe80:ffff:0:0:0:0:0:0]/32", (struct sockaddr_storage *)&addr, &if_index), -1);
-    ASSERT_EQ(aeron_find_interface("127.0.0.10/32", (struct sockaddr_storage *)&addr, &if_index), -1);
-    ASSERT_EQ(aeron_find_interface("172.16.1.20/12", (struct sockaddr_storage *)&addr, &if_index), -1);
-    ASSERT_EQ(aeron_find_interface("192.168.2.20/24", (struct sockaddr_storage *)&addr, &if_index), -1);
 }
